fix: pollResponseTimer throwing exception
PollResponseTimer takes almost 7min to run the first request (pending
polls for the last day), but is on a 5min schedule. The thread is
interrupted, but sometimes fails to clean everything up and release the
jdbc connection. The method throws an exception which is not recommended
based on the documentation.

This commit does not do anything about the query performance.

Refs: FART-547